api/pyproject.toml
[tool.poetry]
name = "divain-api"
version = "0.1.0"
description = "Divain API"
readme = "README.md"
authors = ["Enrique Hidalgo <121901959+ehdlg@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.116.1"
uvicorn = "^0.22.0"
pydantic = "^2.11.7"
sqlmodel = "^0.0.24"
python-dotenv = "^1.1.1"
psycopg2-binary = "^2.9.10"
pytest = "^8.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
